+++ BefunRep/Test/ResultTester.cs
@@ -3,6 +3,6 @@ namespace BefunRep.Test
 {
 	public abstract class ResultTester
 	{
-		public abstract bool test(string code, int result, out string error);
+		public abstract bool test(string code, long result, out string error);
 	}
 }
